refactor: move prase logics to operations
