+++ src/operations.test.ts
@@ -2,7 +2,7 @@ import moment from "moment";
 
 import { Status, StatusType } from "./Status";
 import { Task } from "./Task";
-import { taskOperations } from "./operations";
+import { splitCheckbox, taskOperations } from "./operations";
 
 const DEFAULT_SETTINGS = {
   targetCssClasses: ["checkbox-time-tracker", "ctt"],
@@ -308,5 +308,124 @@ describe("formatTask", () => {
     const result = taskOp.formatTask(task);
 
     expect(result).toBe("  - [x] Task content");
+  });
+});
+
+describe("fromLine", () => {
+  it("- [ ] task content", () => {
+    const body = "- [ ] task content";
+    const taskOp = new taskOperations(DEFAULT_SETTINGS);
+    const result = taskOp.parseLine(body);
+
+    expect(result?.start).toBeUndefined();
+    expect(result?.end).toBeUndefined();
+    expect(result?.taskBody).toBe("task content");
+  });
+
+  it("- [ ] 10:00-12:00(1h30m) task content", () => {
+    const body = "- [ ] task content";
+    const taskOp = new taskOperations(DEFAULT_SETTINGS);
+    const result = taskOp.parseLine(body);
+
+    expect(result?.start).toBeUndefined();
+    expect(result?.end).toBeUndefined();
+    expect(result?.taskBody).toBe("task content");
+  });
+
+  it("- [/] 11:00 10:00-12:00(1h30m) task content", () => {
+    const body = "- [/] 11:00 10:00-12:00(1h30m) task content";
+    const taskOp = new taskOperations(DEFAULT_SETTINGS);
+    const result = taskOp.parseLine(body);
+
+    expect(result?.start?.format("HH:mm")).toBe("11:00");
+    expect(result?.end).toBeUndefined();
+    expect(result?.taskBody).toBe("10:00-12:00(1h30m) task content");
+  });
+
+  it("- [/] 10:00-12:00(1h30m) task content", () => {
+    const body = "- [/] 10:00-12:00(1h30m) task content";
+    expect(() => {
+      const taskOp = new taskOperations(DEFAULT_SETTINGS);
+      taskOp.parseLine(body);
+    }).toThrow("Line does not match Doing regex");
+  });
+
+  it("- [/] 10:00 task content", () => {
+    const body = "- [/] 10:00 task content";
+    const taskOp = new taskOperations(DEFAULT_SETTINGS);
+    const result = taskOp.parseLine(body);
+
+    expect(result?.start?.format("HH:mm")).toBe("10:00");
+    expect(result?.end).toBeUndefined();
+    expect(result?.taskBody).toBe("task content");
+  });
+
+  it("- [/] 10:00 ", () => {
+    const body = "- [/] 10:00 ";
+    const taskOp = new taskOperations(DEFAULT_SETTINGS);
+    const result = taskOp.parseLine(body);
+
+    expect(result?.start?.format("HH:mm")).toBe("10:00");
+    expect(result?.end).toBeUndefined();
+    expect(result?.taskBody).toBe("");
+  });
+
+  it("- [x] 10:00-12:00 task content", () => {
+    const body = "- [x] 10:00-12:00 task content";
+    const taskOp = new taskOperations(DEFAULT_SETTINGS);
+    const result = taskOp.parseLine(body);
+
+    expect(result?.start?.format("HH:mm")).toBe("10:00");
+    expect(result?.end?.format("HH:mm")).toBe("12:00");
+    expect(result?.taskBody).toBe("task content");
+  });
+
+  it("- [x] 10:00-12:00 11:00-12:00(1h) task content", () => {
+    const body = "- [x] 10:00-12:00 11:00-12:00(1h) task content";
+    const taskOp = new taskOperations(DEFAULT_SETTINGS);
+    const result = taskOp.parseLine(body);
+
+    expect(result?.start?.format("HH:mm")).toBe("10:00");
+    expect(result?.end?.format("HH:mm")).toBe("12:00");
+    expect(result?.taskBody).toBe("11:00-12:00(1h) task content");
+  });
+
+  it("- [x] 10:00 - 12:00 11:00-12:00(1h) task content", () => {
+    const body = "- [x] 10:00 - 12:00 11:00-12:00(1h) task content";
+    const taskOp = new taskOperations(DEFAULT_SETTINGS);
+    const result = taskOp.parseLine(body);
+
+    expect(result?.start?.format("HH:mm")).toBe("10:00");
+    expect(result?.end?.format("HH:mm")).toBe("12:00");
+    expect(result?.taskBody).toBe("11:00-12:00(1h) task content");
+  });
+
+  it("- [x] 10:00 task content", () => {
+    const body = "- [x] 10:00 task content";
+
+    expect(() => {
+      const taskOp = new taskOperations(DEFAULT_SETTINGS);
+      taskOp.parseLine(body);
+    }).toThrow("Line does not match Done regex");
+  });
+});
+
+describe("splitCheckbox", () => {
+  it("should split the checkbox line correctly", () => {
+    const line = "  - [x] Task content";
+    const result = splitCheckbox(line);
+
+    expect(result.indentation).toBe("  ");
+    expect(result.listMarker).toBe("-");
+    expect(result.statusSymbol).toBe("x");
+    expect(result.body).toBe("Task content");
+  });
+
+  it("should throw an error if the line does not match the task regex", () => {
+    const line = "Invalid line";
+
+    expect(() => {
+      splitCheckbox(line);
+    }).toThrow("Line does not match task regex");
   });
 });
